# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WaitlistInput {
  email: string;
}

export interface WaitlistEntry {
  id: number;
  email: string;
  createdAt: string;
}

export interface Credentials {
  email: string;
  /** @minLength 10 */
  password: string;
}

export interface Account {
  email: string;
  /** Whether this account may see the growth dashboard. Read from the server's OWNER_EMAIL setting, never stored on the account. */
  isOwner: boolean;
}

export interface Ok {
  ok: boolean;
}

export type StatsTotalsReferralSources = { [key: string]: number };

export interface StatsTotals {
  daily_active: number;
  weekly_active: number;
  checks_today: number;
  checks_7d: number;
  total_clients: number;
  total_registrations: number;
  total_followers: number;
  total_follows: number;
  referral_sources: StatsTotalsReferralSources;
  referrals_answered: number;
}

export interface StatsDailyPoint {
  date: string;
  new_installs: number;
  active_installs: number;
  checks: number;
  new_registrations: number;
}

export interface StatsDaily {
  days: number;
  series: StatsDailyPoint[];
}

export interface StatsProfile {
  display_name?: string | null;
  us_state?: string | null;
  school?: string | null;
  field?: string | null;
  job_search_stage?: string | null;
  visa_status?: string | null;
  graduation?: string | null;
}

export interface StatsRegistrationRow {
  email: string;
  created_at: string;
  installed_at?: string | null;
  verified_at?: string | null;
  referral_source?: string | null;
  subscriptions: number;
  profile?: StatsProfile | null;
}

export interface StatsRegistrations {
  total: number;
  rows: StatsRegistrationRow[];
}

export interface ErrorResponse {
  error: string;
}

export type CoachProblemSummaryDifficulty =
  (typeof CoachProblemSummaryDifficulty)[keyof typeof CoachProblemSummaryDifficulty];

export const CoachProblemSummaryDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface CoachProblemSummary {
  id: string;
  num: number;
  title: string;
  slug: string;
  difficulty: CoachProblemSummaryDifficulty;
  patterns: string[];
  neetcodeGroup: string;
}

export type CoachPlanReviewItemMode =
  (typeof CoachPlanReviewItemMode)[keyof typeof CoachPlanReviewItemMode];

export const CoachPlanReviewItemMode = {
  grill: "grill",
  "re-solve": "re-solve",
} as const;

export type CoachPlanReviewItemGrade =
  | (typeof CoachPlanReviewItemGrade)[keyof typeof CoachPlanReviewItemGrade]
  | null;

export const CoachPlanReviewItemGrade = {
  pass: "pass",
  partial: "partial",
  fail: "fail",
} as const;

export interface CoachPlanReviewItem {
  problem: CoachProblemSummary;
  mode: CoachPlanReviewItemMode;
  minutes: number;
  risk: number;
  daysOverdue: number;
  weakPoints: string[];
  done: boolean;
  solved: boolean;
  grade?: CoachPlanReviewItemGrade;
}

export type CoachPlanNewItemGrade =
  | (typeof CoachPlanNewItemGrade)[keyof typeof CoachPlanNewItemGrade]
  | null;

export const CoachPlanNewItemGrade = {
  pass: "pass",
  partial: "partial",
  fail: "fail",
} as const;

export interface CoachPlanNewItem {
  problem: CoachProblemSummary;
  minutes: number;
  score?: number | null;
  done: boolean;
  solved: boolean;
  grade?: CoachPlanNewItemGrade;
}

export interface CoachPlan {
  date: string;
  budget: number;
  plannedMinutes: number;
  reviews: CoachPlanReviewItem[];
  new: CoachPlanNewItem[];
  deferredReviews: number;
  sprint: boolean;
  sprintDays?: number | null;
  totalSeen: number;
  totalProblems: number;
  doneToday: number;
  solvedToday: number;
  assignedToday: number;
}

export interface CoachSolvedInput {
  problemId: string;
  solved: boolean;
}

export type CoachGradeInputGrade =
  (typeof CoachGradeInputGrade)[keyof typeof CoachGradeInputGrade];

export const CoachGradeInputGrade = {
  pass: "pass",
  partial: "partial",
  fail: "fail",
} as const;

export type CoachGradeInputMode =
  (typeof CoachGradeInputMode)[keyof typeof CoachGradeInputMode];

export const CoachGradeInputMode = {
  grill: "grill",
  "re-solve": "re-solve",
} as const;

export interface CoachGradeInput {
  problemId: string;
  grade: CoachGradeInputGrade;
  /** @maxItems 6 */
  weakPoints?: string[];
  mode?: CoachGradeInputMode;
  /** @maxLength 2000 */
  notes?: string;
}

export type CoachGradeResultState =
  (typeof CoachGradeResultState)[keyof typeof CoachGradeResultState];

export const CoachGradeResultState = {
  new: "new",
  learning: "learning",
  review: "review",
  mastered: "mastered",
} as const;

export interface CoachGradeResult {
  problemId: string;
  state: CoachGradeResultState;
  intervalDays: number;
  due: string;
  lapses: number;
  weakPoints: string[];
}

export interface CoachForecastDay {
  date: string;
  count: number;
  minutes: number;
}

export interface CoachForecast {
  days: CoachForecastDay[];
}

export type CoachDoneEntryGrade =
  (typeof CoachDoneEntryGrade)[keyof typeof CoachDoneEntryGrade];

export const CoachDoneEntryGrade = {
  pass: "pass",
  partial: "partial",
  fail: "fail",
} as const;

export interface CoachDoneEntry {
  id: string;
  grade: CoachDoneEntryGrade;
  mode: string;
}

export type CoachLogDayStatus =
  (typeof CoachLogDayStatus)[keyof typeof CoachLogDayStatus];

export const CoachLogDayStatus = {
  complete: "complete",
  partial: "partial",
  ungraded: "ungraded",
  missed: "missed",
  extra: "extra",
  rest: "rest",
  pending: "pending",
} as const;

export interface CoachLogDay {
  day: string;
  status: CoachLogDayStatus;
  assignedNew: string[];
  assignedReviews: string[];
  plannedMinutes: number;
  solved: string[];
  done: CoachDoneEntry[];
}

export interface CoachAdherence {
  window: number;
  assignedDays: number;
  finishedDays: number;
  workedDays: number;
  rate: number;
}

export interface CoachLog {
  days: CoachLogDay[];
  streak: number;
  adherence: CoachAdherence;
}

export interface CoachConfig {
  dailyMinutes: number;
  newPerDay: number;
  sprintWindowDays: number;
  interviewDate?: string | null;
  targetCompanies: string[];
}

export interface CoachConfigInput {
  /**
   * @minimum 15
   * @maximum 480
   */
  dailyMinutes?: number;
  /**
   * @minimum 0
   * @maximum 10
   */
  newPerDay?: number;
  /**
   * @minimum 1
   * @maximum 60
   */
  sprintWindowDays?: number;
  interviewDate?: string | null;
  /** @maxItems 20 */
  targetCompanies?: string[];
}

export interface CoachToken {
  /** Shown exactly once. The server stores only its hash. */
  token: string;
}

export type GetStatsDailyParams = {
  /**
   * @minimum 1
   * @maximum 365
   */
  days?: number;
};

export type GetStatsRegistrationsParams = {
  /**
   * @minimum 1
   * @maximum 1000
   */
  limit?: number;
};

export type GetCoachForecastParams = {
  /**
   * @minimum 1
   * @maximum 60
   */
  days?: number;
};

export type GetCoachLogParams = {
  /**
   * @minimum 1
   * @maximum 366
   */
  days?: number;
};