# lib/api-zod/src/generated/types/coachAdherence.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoachAdherence {
  window: number;
  assignedDays: number;
  finishedDays: number;
  workedDays: number;
  rate: number;
}