# lib/api-zod/src/generated/types/coachGradeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachGradeInputGrade } from "./coachGradeInputGrade";
import type { CoachGradeInputMode } from "./coachGradeInputMode";

export interface CoachGradeInput {
  problemId: string;
  grade: CoachGradeInputGrade;
  /** @maxItems 6 */
  weakPoints?: string[];
  mode?: CoachGradeInputMode;
  /** @maxLength 2000 */
  notes?: string;
}