# lib/api-zod/src/generated/types/coachPlanNewItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CoachPlanNewItemGrade } from "./coachPlanNewItemGrade";
import type { CoachProblemSummary } from "./coachProblemSummary";

export interface CoachPlanNewItem {
  problem: CoachProblemSummary;
  minutes: number;
  score?: number | null;
  done: boolean;
  solved: boolean;
  grade?: CoachPlanNewItemGrade;
}